Initial Base Model
We now have a working AI. It's success rate is about 75-78%. The structure has been setup for our project. All the data of the initial run has been added. all of the code has been commented, and links to the documentation is also in the code. (NOTE THE .PICKLE FILE CAN NOT BE IN THE GITHUB. We will no longer be able to commit to the repo)